+++ source/default_mode/Symbiont.h
@@ -59,22 +59,22 @@ protected:
     * Purpose: Tracks the number of reproductive events in this symbiont's lineage.
     *
   */
-  unsigned int reproductions = 0;
+  size_t reproductions = 0;
 
   /**
    *
    * Purpose: Tracks the number of tag flips towards partner in this symbiont's lineage.
    *
  */
-  unsigned int towards_partner_count = 0;
+  size_t towards_partner_count = 0;
 
   /**
     *
     * Purpose: Tracks the number of tag flips away from partner in this symbiont's lineage.
     *
   */
-
-  unsigned int from_partner_count = 0;
+  size_t from_partner_count = 0;
+  
   /**
     *
     * Purpose: Represents an instance of random.
@@ -212,7 +212,7 @@ public:
    *
    * Purpose: To set the count of reproductions in this lineage.
    */
-  void SetReproCount(unsigned int _in) { reproductions = _in; }
+  void SetReproCount(size_t _in) { reproductions = _in; }
 
 
   /**
@@ -222,7 +222,7 @@ public:
    *
    * Purpose: To get the count of reproductions in this lineage.
    */
-  unsigned int GetReproCount() { return reproductions; }
+  size_t GetReproCount() { return reproductions; }
 
   /**
    * Input: Set the flips towards a partner counter
@@ -231,7 +231,7 @@ public:
    *
    * Purpose: To set the count of flips towards a partner in this lineage.
    */
-  void SetTowardsPartnerCount(unsigned int _in) { towards_partner_count = _in; }
+  void SetTowardsPartnerCount(size_t _in) { towards_partner_count = _in; }
 
 
   /**
@@ -241,7 +241,7 @@ public:
    *
    * Purpose: To get the count of flips towards a partner in this lineage.
    */
-  unsigned int GetTowardsPartnerCount() { return towards_partner_count; }
+  size_t GetTowardsPartnerCount() { return towards_partner_count; }
 
 
   /**
@@ -251,7 +251,7 @@ public:
    *
    * Purpose: To set the count of flips from a partner in this lineage.
    */
-  void SetFromPartnerCount(unsigned int _in) { from_partner_count = _in; }
+  void SetFromPartnerCount(size_t _in) { from_partner_count = _in; }
 
 
   /**
@@ -261,7 +261,7 @@ public:
    *
    * Purpose: To get the count of flips from a partner in this lineage.
    */
-  unsigned int GetFromPartnerCount() { return from_partner_count; }
+  size_t GetFromPartnerCount() { return from_partner_count; }
 
   /**
    * Input: None
